Fix memory leak (#31)
* chore(deps): Add leak_tracker_flutter_testing to dev_dependencies

This change adds the leak_tracker_flutter_testing package to the dev_dependencies in pubspec.yaml and creates a new test configuration file for enabling leak detection during tests.

* fix(memory_leak): Properly dispose of the controller in CountrySelectorBaseState

Added a dispose method to the CountrySelectorBaseState class to ensure the controller is disposed of correctly, preventing potential memory leaks.

* chore(version): Bump version to 1.0.17

* chore(changelog): Update changelog for version 1.0.17

Add entry for version 1.0.17 to the changelog, documenting the fix for a memory leak in the CountrySelectorBase widget.